Fix exorbitant time taken to merge/collapse RangeSet information (resolves issue #13765).
The previous implementation had accommodated for situations that were
not to be encouraged--e.g. namely the following set of EventRanges:

   [1,4)
   [1,11)
   [4,11)

In the previous implementation, such a set of EventRanges is sorted
(by the weak-ordering criterion), and it would collapse to one
EventRange, namely [1,11).

However, the above situation is ill-defined in the art context since
the collection of events corresponding to a given product/auxiliary
must be unique.  In other words, a RangeSet is not meant to contain
duplicate events.  It is permissible for two _separate_ RangeSets to
contain duplicate events since they are separate entities.  But the
previous situation was trying to support a situation where a RangeSet
was internally inconsistent.