Use root to do upcasts on macOS.
When using macOS-native clang for compilation the internal
implementation of the C++ ABI is not the same as when using
gcc, so use the root facilities to perform up-casting instead
of the more efficient C++ ABI functions.  This does cause
some autoparsing to happen, but the cost is relatively minor
in this particular case.

In the future it might be possible to develop code that uses
the macOS-native clang internal implementation of the C++ ABI
to do the up-casting more efficiently, but for now we do not
have the resources to devote to such an effort.