Require EDProductGetter for non-default-constructed Ptrs.
The Ptr<T>::parentAs() functionality requires an EDProductGetter to
retrieve the parent product.  For Ptrs that are read off disk, this
EDProductGetter is provided via the RefCoreStreamer.

For in-memory Ptrs created via a (Valid)Handle, no EDProductGetter has
been provided in the past as the collection element has already been
available.  This commit provides the EDProductGetter through the
handle so that the parentAs functionality can be uniformly supported
regardless of how the Ptrs are prepared.